moved inquiry type definition
