+++ frontend/src/hooks/useInquiry/Inquiry.d.ts
@@ -1,0 +1,35 @@
+export type InquiryData = {
+  id?: number;
+  createdOn?: Date;
+  lastUpdated?: Date;
+  state: string;
+  inquiry_completion_date?: Date;
+  candidate_first_name: string;
+  candidate_last_name: string;
+  current_position_number: string;
+  current_position_title: string;
+  // current_ministry: Ministry;
+  current_ministry_id: string | null;
+  current_annual_salary: number;
+  current_mccf_classification: string;
+  // experience_level: Experience;
+  experience_level_id: string | null;
+  new_position_number: string;
+  new_position_title: string;
+  // new_mccf_classification: SalaryRange;
+  new_mccf_classification_id: string | null;
+  // appointment: Appointment;
+  appointment_id: string | null;
+  // process: Process;
+  process_id: string | null;
+  salary_estimate: number;
+  hm_comment: string;
+  shr_comment: string;
+  adm_comment: string;
+  // hm_user: User;
+  hm_user_id: string | null;
+  // shr_user: User;
+  shr_user_id: string | null;
+  // adm_user: User;
+  adm_user_id: string | null;
+};
