--- frontend/src/hooks/useInquiry/inquiry.fields.tsx
@@ -1,218 +1,0 @@
-import { IFormField } from '@/components';
-
-export const inquiryFormFields: IFormField[] = [
-  {
-    name: 'id',
-    label: 'id',
-    type: 'text',
-    defaultValue: '',
-    section: 'status',
-    hidden: true,
-    sortOrder: 1,
-  },
-  {
-    name: 'state',
-    label: 'Current State',
-    type: 'status',
-    defaultValue: 'draft',
-    section: 'status',
-    sortOrder: 2,
-  },
-  {
-    name: 'candidate_first_name',
-    label: 'First Name',
-    type: 'text',
-    required: true,
-    defaultValue: '',
-    section: 'employee',
-    sortOrder: 1,
-  },
-  {
-    name: 'candidate_last_name',
-    label: 'Last Name',
-    type: 'text',
-    required: true,
-    defaultValue: '',
-    section: 'employee',
-    sortOrder: 2,
-  },
-  {
-    name: 'current_position_number',
-    label: 'Current Position Number',
-    type: 'text',
-    defaultValue: '',
-    section: 'employee',
-    sortOrder: 6,
-  },
-  {
-    name: 'current_position_title',
-    label: 'Current Position Title',
-    type: 'text',
-    defaultValue: '',
-    section: 'employee',
-    sortOrder: 5,
-  },
-  {
-    name: 'current_ministry_id',
-    label: 'Current Ministry',
-    type: 'select',
-    selectionOptions: {
-      endPoint: 'ministry',
-      labelFieldName: 'ministry_name',
-      valueFieldName: 'id',
-    },
-    defaultValue: '',
-    section: 'employee',
-    sortOrder: 3,
-  },
-  {
-    name: 'current_annual_salary',
-    label: 'Current Annual Salary',
-    type: 'currency',
-    defaultValue: 0,
-    section: 'employee',
-    sortOrder: 7,
-  },
-  {
-    name: 'current_mccf_classification',
-    label: 'Current Classification',
-    type: 'text',
-    defaultValue: '',
-    section: 'employee',
-    sortOrder: 4,
-  },
-  {
-    name: 'experience_level_id',
-    label: 'Experience Level',
-    type: 'select',
-    selectionOptions: {
-      endPoint: 'experience',
-      labelFieldName: 'experience_level',
-      valueFieldName: 'id',
-      sortFieldName: 'experience_level',
-    },
-    defaultValue: 1,
-    section: 'position',
-    sortOrder: 5,
-  },
-  {
-    name: 'new_position_number',
-    label: 'New Position Number',
-    type: 'text',
-    required: true,
-    defaultValue: '',
-    section: 'position',
-    sortOrder: 2,
-  },
-  {
-    name: 'new_position_title',
-    label: 'New Position Title',
-    type: 'text',
-    required: true,
-    defaultValue: '',
-    section: 'position',
-    sortOrder: 1,
-  },
-  {
-    name: 'new_mccf_classification_id',
-    label: 'New Classification',
-    type: 'select',
-    required: true,
-    defaultValue: '',
-    section: 'position',
-    sortOrder: 3,
-    selectionOptions: {
-      endPoint: 'salaryRange',
-      labelFieldName: 'band',
-      valueFieldName: 'id',
-      sortFieldName: 'minimum_salary',
-    },
-  },
-  {
-    name: 'appointment_id',
-    label: 'Appointment Type',
-    type: 'select',
-    selectionOptions: {
-      endPoint: 'appointment',
-      labelFieldName: 'type',
-      valueFieldName: 'id',
-    },
-    defaultValue: 2,
-    section: 'position',
-    sortOrder: 6,
-  },
-  {
-    name: 'process_id',
-    label: 'Process Type',
-    type: 'select',
-    selectionOptions: {
-      endPoint: 'process',
-      labelFieldName: 'type',
-      valueFieldName: 'id',
-    },
-    defaultValue: 1,
-    section: 'position',
-    sortOrder: 7,
-  },
-  {
-    name: 'salary_estimate',
-    label: 'Salary Estimate',
-    type: 'currency',
-    defaultValue: 0,
-    section: 'position',
-    sortOrder: 4,
-  },
-  {
-    name: 'hm_comment',
-    label: 'Hiring Manager Comments',
-    type: 'multiline',
-    defaultValue: '',
-    section: 'comment',
-    sortOrder: 1,
-  },
-  {
-    name: 'hm_user',
-    label: 'HM GUID',
-    type: 'text',
-    defaultValue: null,
-    section: 'status',
-    hidden: true,
-    sortOrder: 1,
-  },
-  {
-    name: 'shr_comment',
-    label: 'Strategic Human Resources Comments',
-    type: 'multiline',
-    defaultValue: '',
-    section: 'comment',
-    sortOrder: 2,
-  },
-  {
-    name: 'shr_user',
-    label: 'SHR GUID',
-    type: 'text',
-    defaultValue: null,
-    section: 'status',
-    hidden: true,
-    sortOrder: 1,
-  },
-  {
-    name: 'adm_comment',
-    label: 'Assistant Deputy Minister Comments',
-    type: 'multiline',
-    defaultValue: '',
-    section: 'comment',
-    sortOrder: 3,
-  },
-  {
-    name: 'adm_user',
-    label: 'ADM GUID',
-    type: 'text',
-    defaultValue: null,
-    section: 'status',
-    hidden: true,
-    sortOrder: 1,
-  },
-];
-
-export default inquiryFormFields;
