+++ frontend/src/types/index.ts
@@ -1,7 +1,6 @@
 export * from './DataFactoryConfig.d';
 export * from './Experience.d';
 export * from './FieldProps.d';
-export * from './Inquiry.d';
 export * from './SalaryData.d';
 export * from './State';
 export * from './User.d';
